# docs/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Livraria API","description":"API REST do sistema de livraria. Documentação completa em /docs (Swagger) ou /redoc.","version":"0.1.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Criar conta","description":"Registra um novo usuário. Retorna 400 se o username já existe.","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Register Api V1 Auth Register Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Autenticar e obter token JWT","description":"Valida credenciais e devolve um token JWT. Use o token no header: `Authorization: Bearer <token>`. Proteção de rotas via token é ativada na Fase 6.","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/books":{"get":{"tags":["books"],"summary":"Listar livros","description":"Retorna todos os livros do catálogo com preço e estoque atual. Endpoint público.","operationId":"list_books_api_v1_books_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookOut"},"type":"array","title":"Response List Books Api V1 Books Get"}}}}}},"post":{"tags":["books"],"summary":"Cadastrar livro","description":"Cria um novo livro no catálogo. Retorna 400 se já existir um livro com o mesmo ID. Proteção de role admin será adicionada em fase futura.","operationId":"create_book_api_v1_books_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookCreateIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/books/{book_id}":{"get":{"tags":["books"],"summary":"Buscar livro por ID","description":"Retorna um único livro pelo seu ID. Retorna 404 se não encontrado. Endpoint público.","operationId":"get_book_api_v1_books__book_id__get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"string","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/carts":{"post":{"tags":["carts"],"summary":"Criar carrinho","description":"Cria um carrinho de compras vazio. Requer autenticação.","operationId":"create_cart_api_v1_carts_post","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartOut"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/carts/{cart_id}":{"get":{"tags":["carts"],"summary":"Buscar carrinho","description":"Retorna o carrinho com todos os seus itens e total calculado. Requer autenticação.","operationId":"get_cart_api_v1_carts__cart_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"cart_id","in":"path","required":true,"schema":{"type":"string","title":"Cart Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/carts/{cart_id}/items":{"post":{"tags":["carts"],"summary":"Adicionar item ao carrinho","description":"Adiciona um livro ao carrinho. Se o livro já existir, incrementa a quantidade. Retorna 400 se o estoque for insuficiente ou o livro não existir. Requer autenticação.","operationId":"add_item_api_v1_carts__cart_id__items_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"cart_id","in":"path","required":true,"schema":{"type":"string","title":"Cart Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartItemAddIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/carts/{cart_id}/items/{book_id}":{"delete":{"tags":["carts"],"summary":"Remover item do carrinho","description":"Remove um livro do carrinho. Se o livro não estiver no carrinho, a operação é silenciosa. Requer autenticação.","operationId":"remove_item_api_v1_carts__cart_id__items__book_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"cart_id","in":"path","required":true,"schema":{"type":"string","title":"Cart Id"}},{"name":"book_id","in":"path","required":true,"schema":{"type":"string","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/orders/preview-discount":{"post":{"tags":["orders"],"summary":"Calcular desconto do cupom","description":"Retorna o total com desconto e o valor economizado sem criar o pedido.","operationId":"preview_discount_api_v1_orders_preview_discount_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreviewDiscountIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreviewDiscountOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/orders/checkout":{"post":{"tags":["orders"],"summary":"Finalizar pedido","description":"Executa o checkout completo: reserva estoque, aplica cupom, registra pedido e pagamento numa transação atômica.","operationId":"checkout_api_v1_orders_checkout_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/health":{"get":{"tags":["infra"],"summary":"Health check","description":"Endpoint de saúde — usado pelo Cloud Run para health checks.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"BookCreateIn":{"properties":{"title":{"type":"string","title":"Title"},"price":{"type":"number","title":"Price"},"stock":{"type":"integer","title":"Stock"},"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"}},"type":"object","required":["title","price","stock"],"title":"BookCreateIn"},"BookOut":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"price":{"type":"number","title":"Price"},"stock":{"type":"integer","title":"Stock"}},"type":"object","required":["id","title","price","stock"],"title":"BookOut"},"CartItemAddIn":{"properties":{"book_id":{"type":"string","title":"Book Id"},"quantity":{"type":"integer","title":"Quantity","default":1}},"type":"object","required":["book_id"],"title":"CartItemAddIn"},"CartItemOut":{"properties":{"book_id":{"type":"string","title":"Book Id"},"title":{"type":"string","title":"Title"},"quantity":{"type":"integer","title":"Quantity"},"unit_price":{"type":"number","title":"Unit Price"},"subtotal":{"type":"number","title":"Subtotal"}},"type":"object","required":["book_id","title","quantity","unit_price","subtotal"],"title":"CartItemOut"},"CartOut":{"properties":{"id":{"type":"string","title":"Id"},"items":{"items":{"$ref":"#/components/schemas/CartItemOut"},"type":"array","title":"Items"},"total":{"type":"number","title":"Total"}},"type":"object","required":["id","items","total"],"title":"CartOut"},"CheckoutIn":{"properties":{"cart_id":{"type":"string","title":"Cart Id"},"payment_method":{"type":"string","title":"Payment Method","default":"pix"},"coupon_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Coupon Code"}},"type":"object","required":["cart_id"],"title":"CheckoutIn"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginIn":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginIn"},"OrderItemOut":{"properties":{"book_id":{"type":"string","title":"Book Id"},"title":{"type":"string","title":"Title"},"quantity":{"type":"integer","title":"Quantity"},"unit_price":{"type":"number","title":"Unit Price"},"subtotal":{"type":"number","title":"Subtotal"}},"type":"object","required":["book_id","title","quantity","unit_price","subtotal"],"title":"OrderItemOut"},"OrderOut":{"properties":{"id":{"type":"string","title":"Id"},"items":{"items":{"$ref":"#/components/schemas/OrderItemOut"},"type":"array","title":"Items"},"total":{"type":"number","title":"Total"},"status":{"type":"string","title":"Status"},"coupon_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Coupon Code"},"payment":{"$ref":"#/components/schemas/PaymentOut"}},"type":"object","required":["id","items","total","status","coupon_code","payment"],"title":"OrderOut"},"PaymentOut":{"properties":{"order_id":{"type":"string","title":"Order Id"},"amount":{"type":"number","title":"Amount"},"method":{"type":"string","title":"Method"},"status":{"type":"string","title":"Status"}},"type":"object","required":["order_id","amount","method","status"],"title":"PaymentOut"},"PreviewDiscountIn":{"properties":{"cart_total":{"type":"number","title":"Cart Total"},"coupon_code":{"type":"string","title":"Coupon Code"}},"type":"object","required":["cart_total","coupon_code"],"title":"PreviewDiscountIn"},"PreviewDiscountOut":{"properties":{"discounted_total":{"type":"number","title":"Discounted Total"},"savings":{"type":"number","title":"Savings"}},"type":"object","required":["discounted_total","savings"],"title":"PreviewDiscountOut"},"RegisterIn":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"RegisterIn"},"TokenOut":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}